+++ src/main/kotlin/model/Nutrition.kt
@@ -1,4 +1,4 @@
-package models
+package model
 
 data class Nutrition (
     val calories:Float?,
